app/server/requirements.txt
Flask==3.1.3

psycopg2-binary

prometheus-client